# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "go-portforward",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/cloverzrg/go-portforward",
            "email": "cloverzrg@gmail.com"
        },
        "license": {
            "name": "go-portforward"
        },
        "version": "1.0"
    },
    "paths": {
        "/v1/forward/": {
            "get": {
                "description": "\"\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "network"
                ],
                "summary": "get forward list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页码，默认1",
                        "name": "page_num",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页面大小，默认20",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.DataResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ForwardList"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "\"\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "network"
                ],
                "summary": "add a forward and start it",
                "parameters": [
                    {
                        "description": "请求json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddForward"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.DataResp"
                        }
                    }
                }
            }
        },
        "/v1/forward/{id}": {
            "get": {
                "description": "\"\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "network"
                ],
                "summary": "get froward detail by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.DataResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ForwardDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/forward/{id}/delete": {
            "post": {
                "description": "\"\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "network"
                ],
                "summary": "delete forward by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.DataResp"
                        }
                    }
                }
            }
        },
        "/v1/forward/{id}/start": {
            "post": {
                "description": "\"\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "network"
                ],
                "summary": "start forward by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.DataResp"
                        }
                    }
                }
            }
        },
        "/v1/forward/{id}/stop": {
            "post": {
                "description": "\"\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "network"
                ],
                "summary": "stop forward by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.DataResp"
                        }
                    }
                }
            }
        },
        "/v1/network/interfaces": {
            "get": {
                "description": "\"\"",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "network"
                ],
                "summary": "get network interface list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.DataResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.NetworkInterface"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AddForward": {
            "type": "object",
            "properties": {
                "listen_address": {
                    "type": "string"
                },
                "listen_port": {
                    "type": "integer"
                },
                "network": {
                    "type": "string"
                },
                "target_address": {
                    "type": "string"
                },
                "target_port": {
                    "type": "integer"
                }
            }
        },
        "dto.ForwardDetail": {
            "type": "object",
            "properties": {
                "listen_address": {
                    "type": "string"
                },
                "listen_port": {
                    "type": "integer"
                },
                "network": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "target_address": {
                    "type": "string"
                },
                "target_port": {
                    "type": "integer"
                }
            }
        },
        "dto.ForwardList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ForwardDetail"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.NetworkInterface": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "default_gateway": {
                    "type": "boolean"
                },
                "desc": {
                    "type": "string"
                }
            }
        },
        "resp.DataResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}